+++ package.json
@@ -10,7 +10,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@nadjitan/surveyor": "^0.7.12",
+    "@nadjitan/surveyor": "^0.8.17",
     "daisyui": "^2.17.0",
     "next": "12.1.6",
     "react": "18.2.0",
