+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import { initSurveyor } from "@nadjitan/surveyor"
 import Head from "next/head"
 import { ReactElement, useEffect } from "react"
-import MainLayout from "../components/layouts/MainLayout"
+import MainLayout from "../components/layouts/main-layout"
 import { NextPageWithLayout } from "./_app"
 import React from "react"
 import Link from "next/link"
