+++ pages/items.tsx
@@ -2,7 +2,7 @@ import { initSurveyor } from "@nadjitan/surveyor"
 import Head from "next/head"
 import Link from "next/link"
 import { ReactElement, useEffect } from "react"
-import MainLayout from "../components/layouts/MainLayout"
+import MainLayout from "../components/layouts/main-layout"
 import { NextPageWithLayout } from "./_app"
 
 const Items: NextPageWithLayout = () => {
