+++ styles/globals.css
@@ -31,7 +31,7 @@
   html,
   body,
   #__next {
-    @apply h-full w-full;
+    @apply min-h-screen w-full;
   }
 }
